refactor(View,Utils): 클래스 이름 개선 및 패키지 분리
