+++ src/main/java/step2/StringUtils.java
@@ -2,7 +2,7 @@ package step2;
 
 import java.util.Objects;
 
-public class Utils {
+public class StringUtils {
     public static boolean isNullOrEmpty(String string) {
         return Objects.isNull(string) || string.trim().isEmpty();
     }
