+++ src/main/java/step2/view/OutputView.java
@@ -1,0 +1,7 @@
+package step2.view;
+
+public class OutputView {
+    public static void printResult(int result) {
+        System.out.println("결과: " + result);
+    }
+}
