+++ src/test/java/step2/domain/CalculatorTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,10 +6,11 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import step2.domain.Calculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static step2.Calculator.NULL_OR_EMPTY_STRING_EXCEPTION_MESSAGE;
+import static step2.domain.Calculator.NULL_OR_EMPTY_STRING_EXCEPTION_MESSAGE;
 
 public class CalculatorTest {
